Implemented checkokut page, starting refactoring
