+++ extractedTranslations/ru/translation.json
@@ -32,6 +32,7 @@
   "Контакты": "Контакты",
   "Корзина": "Корзина",
   "Краткое содержание": "Краткое содержание",
+  "Купить за": "Купить за",
   "Логин": "Логин",
   "Новостная рассылка": "Новостная рассылка",
   "Номер телефона": "Номер телефона",
