+++ src/pages-components/CheckoutPage/ui/CheckoutPage/ProductsPanel/Goods/Goods.module.scss
