+++ src/pages-components/CheckoutPage/ui/CheckoutPage/ProductsPanel/Goods/GoodsItem/GoodsItem.module.scss
