+++ src/pages-components/CheckoutPage/ui/CheckoutPage/ProductsPanel/Summary/Summary.module.scss
