+++ src/pages-components/CheckoutPage/ui/CheckoutPage/UserPanel/Payment/Payment.module.scss
@@ -1,0 +1,130 @@
+.Payment {
+  display: flex;
+  flex-direction: column;
+  gap: 30px;
+  margin-bottom: 40px;
+}
+
+.TitleWrapper {
+  display: flex;
+  align-items: center;
+  gap: 20px;
+}
+
+.Title {
+  font-size: 24px;
+  font-weight: 500;
+  line-height: 30px;
+  letter-spacing: -1px;
+}
+
+.Stick {
+  height: 1px;
+  width: 100%;
+  background-color: var(--primary-dark-5);
+}
+
+.InputsList {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  justify-content: space-between;
+  row-gap: 30px;
+}
+
+.InputWrapper {
+  max-width: 50%;
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+
+.Label {
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 18px;
+  letter-spacing: -0.5px;
+  color: var(--label-color);
+  cursor: pointer;
+  transition: all .25s ease-in-out;
+  &:hover {
+    color: var(--accent-blue-2);
+  }
+}
+
+.PaymentInputs {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+
+.PaymentMethodIcons {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  gap: 20px;
+}
+
+.PaymentIcon {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 50px;
+  width: 70px;
+  border: 1px solid var(--primary-dark-4);
+  border-radius: 3px;
+  cursor: pointer;
+  transition: all .25s ease-in-out;
+  &:hover {
+    border: 1px solid var(--accent-blue-1);
+    box-shadow: rgba(99, 99, 99, 0.2) 0px 2px 8px 0px;
+  }
+}
+
+.FieldWrapper {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  padding: 15px;
+  border-radius: 10px;
+  border: 1px solid var(--primary-dark-5);
+  width: 360px;
+  cursor: pointer;
+  transition: all .25s ease-in-out;
+  &:focus-within {
+    border: 1px solid var(--accent-blue-2);
+  }
+}
+
+.Input {
+  appearance: none;
+  flex: 1;
+  &::placeholder {
+    font-size: 18px;
+    font-weight: 400;
+    line-height: 20px;
+    letter-spacing: -0.3px;
+    color: var(--label-color);
+  }
+}
+
+.PurchaseButton {
+  padding: 18px 20px;
+  width: 212px;
+  height: 60px;
+  background-color: var(--accent-blue-2);
+  color: var(--primary-light-1);
+  display: flex;
+  align-items: center;
+  border-radius: 10px;
+  gap: 7px;
+  font-size: 20px;
+  font-weight: 500;
+  line-height: 24px;
+  letter-spacing: -0.6px;
+
+  transition: all .25s ease-in-out;
+  &:hover {
+    box-shadow: rgba(50, 50, 93, 0.25) 0px 30px 60px -12px, rgba(0, 0, 0, 0.3) 0px 18px 36px -18px;
+  }
+}
