+++ src/pages-components/CheckoutPage/ui/CheckoutPage/UserPanel/UserPanel.tsx
@@ -1,9 +1,9 @@
 import { FC } from 'react';
 import { classNames } from 'shared/lib/classNames/classNames';
 import cls from './UserPanel.module.scss';
-import Contacts from '../Contacts/Contacts';
-import Address from '../Address/Address';
-import Payment from '../Payment/Payment';
+import Contacts from './Contacts/Contacts';
+import Address from './Address/Address';
+import Payment from './Payment/Payment';
 
 interface UserPanelProps {
   className?: string;
