--- src/pages-components/CheckoutPage/ui/Payment/Payment.module.scss
@@ -1,66 +1,0 @@
-.Payment {
-  display: flex;
-  flex-direction: column;
-  gap: 30px;
-  margin-bottom: 40px;
-}
-
-.TitleWrapper {
-  display: flex;
-  align-items: center;
-  gap: 20px;
-}
-
-.Title {
-  font-size: 24px;
-  font-weight: 500;
-  line-height: 30px;
-  letter-spacing: -1px;
-}
-
-.Stick {
-  height: 1px;
-  width: 100%;
-  background-color: var(--primary-dark-5);
-}
-
-.InputsList {
-  display: flex;
-  flex-wrap: wrap;
-  width: 100%;
-  justify-content: space-between;
-  row-gap: 30px;
-}
-
-.InputWrapper {
-  max-width: 50%;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.Label {
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 18px;
-  letter-spacing: -0.5px;
-  color: var(--label-color);
-  cursor: pointer;
-  transition: all .25s ease-in-out;
-  &:hover {
-    color: var(--accent-blue-2);
-  }
-}
-
-.Input {
-  appearance: none;
-  padding: 15px;
-  border-radius: 10px;
-  border: 1px solid var(--primary-dark-5);
-  width: 275px;
-  cursor: pointer;
-  transition: all .25s ease-in-out;
-  &:focus-within {
-    border: 1px solid var(--accent-blue-2);
-  }
-}
