+++ src/pages-components/ProductsPage/ui/ProductsPage/CategoryFilters/CategoryFilters.module.scss
