+++ src/pages-components/ProductsPage/ui/ProductsPage/CollectionFilters/CollectionFilters.module.scss
