+++ src/pages-components/ProductsPage/ui/ProductsPage/FilterItem/FilterItem.module.scss
