+++ src/pages-components/ProductsPage/ui/ProductsPage/FilterItem/FilterItem.tsx
