+++ src/pages-components/ProductsPage/ui/ProductsPage/PriceFilters/PriceFilters.module.scss
