+++ src/pages-components/ProductsPage/ui/ProductsPage/PriceFilters/PriceFilters.tsx
