+++ src/pages-components/ProductsPage/ui/ProductsPage/ProductsDropdown/ProductsDropdown.module.scss
