+++ src/pages-components/ProductsPage/ui/ProductsPage/StyleFilters/StyleFilters.module.scss
